🔨 : tetrisSocket.ts 리팩토링
- 함수 파일 분리
